pyproject.toml
[project]
name = "kannada-news-automation"
version = "0.1.0"
description = "Automated Kannada news shorts generation with ElevenLabs TTS"
authors = [{name = "News Automation", email = "admin@example.com"}]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # LlamaIndex core
    "llama-index>=0.10.0",
    "llama-index-llms-openai",
    "llama-index-embeddings-openai", 
    "llama-index-vector-stores-chroma",
    "chromadb>=0.4.0",
    
    # Web framework and async
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "celery[redis]>=5.3.0",
    "aiohttp>=3.9.0",
    
    # Media processing
    "moviepy>=1.0.3",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "pydub>=0.25.1",
    
    # Text-to-Speech
    "elevenlabs>=0.2.26",
    
    # APIs and utilities
    "requests>=2.31.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "schedule>=1.2.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true